requirements.txt
streamlit
google-generativeai
PyMuPDF
pillow
pandas
python-docx